enunciado03/requirements.txt
PyGithub>=2.3.0
pandas>=2.2.0
requests>=2.32.0
python-dotenv>=1.0.0
scipy>=1.14.0
numpy>=2.0.0
gql>=3.5.0
aiohttp>=3.10.0
